# apigateway/openapi.yaml
openapi: 3.1.0
info:
  title: E-commerce API Gateway
  version: 1.1.0
  description: Consolidated OpenAPI spec for the API Gateway proxying User, Product, and Order services.
servers:
  - url: http://localhost:8083
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        code: { type: string }
        details: { type: object, additionalProperties: true }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string, format: email }
        token: { type: string, description: Bearer JWT token }
    Address:
      type: object
      properties:
        id: { type: string }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        state: { type: string }
        postal_code: { type: string }
        country: { type: string }
        phone: { type: string, nullable: true }
        is_default: { type: boolean }
    CreateAddressRequest:
      type: object
      required: [line1, city, state, postal_code, country]
      properties:
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        postal_code: { type: string }
        country: { type: string, minLength: 2, maxLength: 2 }
        phone: { type: string }
        is_default: { type: boolean }
    UpdateAddressRequest:
      type: object
      properties:
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        postal_code: { type: string }
        country: { type: string }
        phone: { type: string }
        is_default: { type: boolean }
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
    CreateUserRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }
        phone: { type: string }
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # Users
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/users/{userId}:
    get:
      tags: [Users]
      summary: Get user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/users/{userId}/addresses:
    get:
      tags: [Addresses]
      summary: List addresses
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Address' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    post:
      tags: [Addresses]
      summary: Create address
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAddressRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/users/{userId}/addresses/{addressId}:
    put:
      tags: [Addresses]
      summary: Update address
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
        - in: path
          name: addressId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAddressRequest' }
      responses:
        '200': { description: Updated }
        '404':
          description: Address not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Addresses]
      summary: Delete address
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
        - in: path
          name: addressId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deleted }
        '404':
          description: Address not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # Products
...